Refactor: 날짜 필터를 YEAR/MONTH 함수에서 범위 비교로 변경해 인덱스 활용

+++ core-api/src/main/java/com/pocketstock/core/budget/dto/AutoBudgetGoalResponse.java
@@ -1,5 +1,6 @@
 package com.pocketstock.core.budget.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record AutoBudgetGoalResponse(Long monthlyBudget, List<BudgetGoalItem> categories) {}
+public record AutoBudgetGoalResponse(BigDecimal monthlyBudget, List<BudgetGoalItem> categories) {}
